chore: move ExternalLibraryAccessor
